HW 3 3 Implement tests for meal service

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcMealRepository.java
@@ -60,7 +60,7 @@ public class JdbcMealRepository implements MealRepository {
 
     @Override
     public boolean delete(int id, int userId) {
-        return jdbcTemplate.update("DELETE FROM user_meals WHERE id=?", id) != 0;
+        return jdbcTemplate.update("DELETE FROM user_meals WHERE user_id=? AND id=?", userId, id) != 0;
     }
 
     @Override
